jawhar

+++ owner-dashboard/src/components/Chat/Messages.jsx
@@ -27,6 +27,7 @@ const Messages = () => {
 
 
     const getConvos = async () => {
+        console.log('gettingConvos')
         try {
             const { data } = await axios.get('http://localhost:3000/api/messages/owner/conversations')
             const uniqueConversationsMap = {};
@@ -84,41 +85,49 @@ const Messages = () => {
 
     const handleChange = (e) => {
         setNewMessage(e.target.value)
-        console.log(newMessage)
     }
 
     const handleSubmit = async (e) => {
         e.preventDefault()
 
-        socket.current.emit('sendMessage', {
-            receiverId: currentChat,
-            text: newMessage
-        })
+        let toSend = newMessage
 
+        if (newMessage.length > 0 && newMessage !== "\n") {
 
-        try {
-
-            const { data } = await axios.post(`http://localhost:3000/api/messages/owner/${currentChat}`, { message: newMessage })
-            setMessages([...messages, data])
-            setNewMessage('')
-
-
-
-        } catch (error) {
-            console.log(error)
-            if (
-                error.response &&
-                error.response.status === 403 &&
-                error.response.data.message === "Owner account unauthorized"
-            ) {
-                toast.error("Upgrade to premium to access our chat feature now!");
-                navigate('/home')
-
+            if (newMessage.slice(-1) === "\n") {
+                toSend = newMessage.slice(0, -1)
             }
 
-            else if (error.response.status === 403 || error.response.status === 401) {
-                localStorage.clear()
-                navigate('/')
+            socket.current.emit('sendMessage', {
+                receiverId: currentChat,
+                text: toSend
+            })
+
+
+            try {
+
+                const { data } = await axios.post(`http://localhost:3000/api/messages/owner/${currentChat}`, { message: toSend })
+                setMessages([...messages, data])
+                setNewMessage('')
+
+
+
+            } catch (error) {
+                console.log(error)
+                if (
+                    error.response &&
+                    error.response.status === 403 &&
+                    error.response.data.message === "Owner account unauthorized"
+                ) {
+                    toast.error("Upgrade to premium to access our chat feature now!");
+                    navigate('/home')
+
+                }
+
+                else if (error.response.status === 403 || error.response.status === 401) {
+                    localStorage.clear()
+                    navigate('/')
+                }
             }
         }
     }
@@ -130,6 +139,21 @@ const Messages = () => {
 
         }
     }
+    const removeMessageNotificationBadge = async () => {
+        try {
+            const { data } = await axios.put(
+                `http://localhost:3000/api/messages/owner/notification`
+            );
+            dispatch(setMessageNotificationBadge(data));
+        } catch (error) {
+            console.log(error);
+            if (error.response.status === 403 || error.response.status === 401) {
+                localStorage.clear();
+                navigate("/");
+            }
+        }
+    };
+
 
     useEffect(() => {
         getMessages()
@@ -152,7 +176,9 @@ const Messages = () => {
     }, [messages])
 
 
+
     useEffect(() => {
+        removeMessageNotificationBadge()
         if (!socket.current) {
             socket.current = io("ws://localhost:8900", {
                 auth: {
@@ -194,10 +220,10 @@ const Messages = () => {
     return (
         <div>
             <NavBar />
-            <div className="messenger">
+            <div className="messenger ">
 
-                <div className="chatMenu"></div>
-                <div className="chatMenuWrapper">
+                <div className="chatMenu  "></div>
+                <div className="chatMenuWrapper ">
                     {conversations.map((convo) =>
                         <div onClick={() => dispatch(setCurrentChat(convo.customerId))} key={convo.customerId}>
                             <Conversations customerId={convo.customerId} />
